refactor: jwp와 tomcat 간 패키지 분리 리팩토링
