--- tomcat/src/main/java/nextstep/jwp/commandcontroller/Controller.java
@@ -1,10 +1,0 @@
-package nextstep.jwp.commandcontroller;
-
-import nextstep.jwp.request.HttpRequest;
-import nextstep.jwp.response.HttpResponse;
-
-public interface Controller {
-
-    boolean canHandle(HttpRequest httpRequest);
-    void service(HttpRequest httpRequest, HttpResponse httpResponse) throws Exception;
-}
