--- tomcat/src/main/java/nextstep/jwp/commandcontroller/MainPageController.java
@@ -1,25 +1,0 @@
-package nextstep.jwp.commandcontroller;
-
-import nextstep.jwp.common.ContentType;
-import nextstep.jwp.common.HttpMethod;
-import nextstep.jwp.common.HttpStatus;
-import nextstep.jwp.request.HttpRequest;
-import nextstep.jwp.response.HttpResponse;
-import nextstep.jwp.response.StatusLine;
-
-public class MainPageController extends AbstractController {
-
-    @Override
-    public boolean canHandle(HttpRequest request) {
-        final HttpMethod method = request.getHttpMethod();
-        final String uri = request.getRequestUri();
-        return method.equals(HttpMethod.GET) && uri.isEmpty();
-    }
-
-    @Override
-    public void service(HttpRequest request, HttpResponse response) {
-        response.setStatusLine(StatusLine.of(request.getHttpVersion(), HttpStatus.OK));
-        response.addHeader("Content-Type", ContentType.HTML.getType());
-        response.setResponseBody("Hello world!");
-    }
-}
