--- tomcat/src/main/java/nextstep/jwp/common/HttpMethod.java
@@ -1,16 +1,0 @@
-package nextstep.jwp.common;
-
-import java.util.Arrays;
-import nextstep.jwp.exception.InvalidHttpMethodException;
-
-public enum HttpMethod {
-    GET,
-    POST;
-
-    public static HttpMethod from(final String method) {
-        return Arrays.stream(values())
-                .filter(httpMethod -> httpMethod.name().equals(method))
-                .findAny()
-                .orElseThrow(InvalidHttpMethodException::new);
-    }
-}
