--- tomcat/src/main/java/nextstep/jwp/common/HttpStatus.java
@@ -1,18 +1,0 @@
-package nextstep.jwp.common;
-
-public enum HttpStatus {
-    OK(200),
-    FOUND(302),
-    UNAUTHORIZED(401),
-    NOT_FOUND(404);
-
-    private final int statusCode;
-
-    HttpStatus(final int statusCode) {
-        this.statusCode = statusCode;
-    }
-
-    public int getStatusCode() {
-        return statusCode;
-    }
-}
