--- tomcat/src/main/java/nextstep/jwp/exception/InvalidHttpMethodException.java
@@ -1,8 +1,0 @@
-package nextstep.jwp.exception;
-
-public class InvalidHttpMethodException extends RuntimeException {
-
-    public InvalidHttpMethodException() {
-        super("잘못된 Http Method입니다.");
-    }
-}
