--- tomcat/src/main/java/nextstep/jwp/exception/UnSupportedContentTypeException.java
@@ -1,8 +1,0 @@
-package nextstep.jwp.exception;
-
-public class UnSupportedContentTypeException extends RuntimeException {
-
-    public UnSupportedContentTypeException() {
-        super("지원하지 않는 Content-Type 입니다.");
-    }
-}
