+++ tomcat/src/main/java/org/apache/catalina/Controller.java
@@ -1,0 +1,9 @@
+package org.apache.catalina;
+
+import org.apache.coyote.request.HttpRequest;
+import org.apache.coyote.response.HttpResponse;
+
+public interface Controller {
+
+    void service(HttpRequest httpRequest, HttpResponse httpResponse) throws Exception;
+}
