+++ tomcat/src/main/java/org/apache/coyote/common/Session.java
@@ -1,0 +1,30 @@
+package org.apache.coyote.common;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class Session {
+
+    private final String id;
+    private final Map<String, Object> values = new HashMap<>();
+
+    public Session(final String id) {
+        this.id = id;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public Object getAttribute(final String name) {
+        return values.get(name);
+    }
+
+    public void setAttribute(final String name, final Object value) {
+        values.put(name, value);
+    }
+
+    public void removeAttribute(final String name) {
+        values.remove(name);
+    }
+}
