+++ tomcat/src/main/java/org/apache/coyote/exception/InvalidHttpMethodException.java
@@ -1,0 +1,8 @@
+package org.apache.coyote.exception;
+
+public class InvalidHttpMethodException extends RuntimeException {
+
+    public InvalidHttpMethodException() {
+        super("잘못된 Http Method입니다.");
+    }
+}
