+++ tomcat/src/main/java/org/apache/coyote/exception/InvalidHttpVersionException.java
@@ -1,0 +1,8 @@
+package org.apache.coyote.exception;
+
+public class InvalidHttpVersionException extends RuntimeException {
+
+    public InvalidHttpVersionException() {
+        super("지원하지 않는 Http 버전입니다.");
+    }
+}
