+++ tomcat/src/main/java/org/apache/coyote/exception/InvalidQueryStringException.java
@@ -1,0 +1,8 @@
+package org.apache.coyote.exception;
+
+public class InvalidQueryStringException extends RuntimeException {
+
+    public InvalidQueryStringException() {
+        super("쿼리 스트링은 key=value 형식이어야 합니다.");
+    }
+}
