+++ tomcat/src/main/java/org/apache/coyote/exception/UnSupportedContentTypeException.java
@@ -1,0 +1,8 @@
+package org.apache.coyote.exception;
+
+public class UnSupportedContentTypeException extends RuntimeException {
+
+    public UnSupportedContentTypeException() {
+        super("지원하지 않는 Content-Type 입니다.");
+    }
+}
