+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -3,19 +3,11 @@ package org.apache.coyote.http11;
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
 import java.net.Socket;
-import java.util.List;
-import nextstep.jwp.RequestMapping;
-import nextstep.jwp.commandcontroller.Controller;
-import nextstep.jwp.commandcontroller.ErrorPageController;
-import nextstep.jwp.commandcontroller.LoginController;
-import nextstep.jwp.commandcontroller.LoginPageController;
-import nextstep.jwp.commandcontroller.MainPageController;
-import nextstep.jwp.commandcontroller.RegisterController;
-import nextstep.jwp.commandcontroller.RegisterPageController;
-import nextstep.jwp.commandcontroller.StaticResourceController;
-import nextstep.jwp.request.HttpRequest;
-import nextstep.jwp.response.HttpResponse;
+import nextstep.jwp.controller.RequestMapping;
+import org.apache.catalina.Controller;
 import org.apache.coyote.Processor;
+import org.apache.coyote.request.HttpRequest;
+import org.apache.coyote.response.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -23,11 +15,10 @@ public class Http11Processor implements Runnable, Processor {
 
     private static final Logger log = LoggerFactory.getLogger(Http11Processor.class);
 
-    private final RequestMapping requestMapping;
+    private final Controller requestMapping = RequestMapping.getDefault();
     private final Socket connection;
 
     public Http11Processor(final Socket connection) {
-        this.requestMapping = registerControllers();
         this.connection = connection;
     }
 
@@ -43,25 +34,11 @@ public class Http11Processor implements Runnable, Processor {
              final var outputStream = connection.getOutputStream()) {
             final HttpRequest httpRequest = HttpRequest.from(bufferedReader);
             final HttpResponse httpResponse = new HttpResponse();
-            final Controller controller = requestMapping.getController(httpRequest);
-            controller.service(httpRequest, httpResponse);
+            requestMapping.service(httpRequest, httpResponse);
             outputStream.write(httpResponse.toResponse().getBytes());
             outputStream.flush();
         } catch (Exception e) {
             log.error(e.getMessage(), e);
         }
-    }
-
-    private RequestMapping registerControllers() {
-        final RequestMapping mapper = new RequestMapping(List.of(
-                new LoginController(),
-                new LoginPageController(),
-                new RegisterController(),
-                new RegisterPageController(),
-                new MainPageController(),
-                new StaticResourceController()
-        ));
-        mapper.setDefaultController(new ErrorPageController());
-        return mapper;
     }
 }
