+++ tomcat/src/main/java/org/apache/coyote/request/RequestLine.java
@@ -1,0 +1,46 @@
+package org.apache.coyote.request;
+
+import java.util.Map;
+import org.apache.coyote.common.HttpMethod;
+import org.apache.coyote.common.HttpVersion;
+
+public class RequestLine {
+
+    private static final int METHOD_INDEX = 0;
+    private static final int URI_INDEX = 1;
+    private static final int VERSION_INDEX = 2;
+
+    private final HttpMethod httpMethod;
+    private final RequestUri requestUri;
+    private final HttpVersion httpVersion;
+
+    private RequestLine(HttpMethod httpMethod, RequestUri requestURI, HttpVersion httpVersion) {
+        this.httpMethod = httpMethod;
+        this.requestUri = requestURI;
+        this.httpVersion = httpVersion;
+    }
+
+    public static RequestLine from(final String requestLine) {
+        final String[] params = requestLine.split(" ");
+        final HttpMethod httpMethod = HttpMethod.from(params[METHOD_INDEX]);
+        final RequestUri uri = RequestUri.from(params[URI_INDEX].substring(1));
+        final HttpVersion httpVersion = HttpVersion.from(params[VERSION_INDEX]);
+        return new RequestLine(httpMethod, uri, httpVersion);
+    }
+
+    public HttpMethod getHttpMethod() {
+        return httpMethod;
+    }
+
+    public String getRequestUri() {
+        return requestUri.getUri();
+    }
+
+    public Map<String, String> getQueryParams() {
+        return requestUri.getQueryParams();
+    }
+
+    public HttpVersion getHttpVersion() {
+        return httpVersion;
+    }
+}
